refact: utils 파일 type 분리
