+++ types/region.ts
@@ -1,0 +1,11 @@
+export interface RegionResponseData {
+  regcodes: {
+    code: string;
+    name: string;
+  }[];
+}
+
+export interface RegionData {
+  value: string;
+  label: string;
+}
