+++ types/validate.ts
@@ -1,0 +1,4 @@
+export interface FieldErrorMessage {
+  keys: string[];
+  label: string;
+}
